requirements.txt
# Core dependencies
abc-radio-wrapper==0.2.0
Pillow>=10.0.0
requests>=2.31.0
tzdata>=2023.3  # Required for timezone support on Windows
